Fixed wrong comment about recharge time and increase BAT2 Shield Cooldown a bit
I spent about 2 hours today to dig behind Shield generators again. Couple of things changed since implementation.
Important to know is that it always take 5-6 seconds until the shield is going to charge up again. There is no setting for that.
The Cooldown in combination with the blinking red 10% shield value is also not always accurate in the game. Especially in MP.
So overall I felt that the BA T2 Cooldown is good enough with 15 Seconds instead of 10. 15 Seconds is the vanilla value, maybe because of the inaccurate handling of timers.